app/requirements.txt
flask==2.0.1
pyautogui==0.9.53
python-dotenv==0.19.0
cryptography==41.0.7
google-api-python-client==2.86.0
PyQt5==5.15.9
qrcode==7.4.2
Pillow==10.0.0
